// src/app/interfaces/carpeta/suplidos.ts
export interface Suplidos {
    TEXTOPANTALLA: string;
    FORMULAEUROS: string;
    DESCRIPCIONCAMPODATOS: string;
    ALIAS: string;
    CODIGOCAMPODATOS: string;
    TIPODATO: string;
    PORCENTAJEIVA: string;
    PORCENTAJERETENIBLE: string;
    REPARTOSPREDEFINIDOS: string;
    MENSAJEAYUDA: string;
    VALORNUMERICO: string;
    VALORTEXTO: string;
    MANUALOCALCULADO?: string;
}